+++ backend/store/etcd/result_filters.go
@@ -1,11 +1,14 @@
 package etcd
 
-import "github.com/sensu/sensu-go/types"
+import (
+	"github.com/sensu/sensu-go/backend/store"
+	"github.com/sensu/sensu-go/types"
+)
 
 // rejectByEnvironment returns false if any given record is not part of the
 // environment configured in the namespace. This method is useful when a
 // wildcard was given for the organization but not for the environment.
-func rejectByEnvironment(ns namespace) func(types.MultitenantResource) bool {
+func rejectByEnvironment(ns store.Namespace) func(types.MultitenantResource) bool {
 	// If the namespace is a wildcard never reject
 	if ns.EnvIsWildcard() {
 		return func(_ types.MultitenantResource) bool {
@@ -14,6 +17,6 @@ func rejectByEnvironment(ns namespace) func(types.MultitenantResource) bool {
 	}
 
 	return func(r types.MultitenantResource) bool {
-		return r.GetEnvironment() != ns.env
+		return r.GetEnvironment() != ns.Env
 	}
 }