+++ backend/store/key_builder_test.go
@@ -1,4 +1,4 @@
-package etcd
+package store
 
 import (
 	"context"
@@ -11,8 +11,8 @@ import (
 func TestSimpleKeyBuilder(t *testing.T) {
 	t.Parallel()
 
-	builder := newKeyBuilder("checks")
-	assert.Equal(t, "/sensu.io/checks", builder.build(""))
+	builder := NewKeyBuilder("checks")
+	assert.Equal(t, "/sensu.io/checks", builder.Build(""))
 }
 
 func TestContextKeyBuilder(t *testing.T) {
@@ -21,6 +21,6 @@ func TestContextKeyBuilder(t *testing.T) {
 	ctx := context.Background()
 	ctx = context.WithValue(ctx, types.OrganizationKey, "org")
 	ctx = context.WithValue(ctx, types.EnvironmentKey, "env")
-	builder := newKeyBuilder("checks").withContext(ctx)
-	assert.Equal(t, "/sensu.io/checks/org/env/check_name", builder.build("check_name"))
+	builder := NewKeyBuilder("checks").WithContext(ctx)
+	assert.Equal(t, "/sensu.io/checks/org/env/check_name", builder.Build("check_name"))
 }